Move files from service directory to services directory (and remove service directory)
